backend(telemetry): fix ESM import path and encapsulate telemetry storage
