+++ apps/backend/routes/usage.ts
@@ -1,13 +1,3 @@
-import { createTelemetryRouter } from '@workbuoy/backend-telemetry';
-import { getTelemetryFallbackStore, ensureTelemetryPersistentStore, isTelemetryPersistenceEnabled } from '../src/telemetryContext.js';
+import { telemetryRouter } from '../src/telemetry/context.js';
 
-const usePersistence = isTelemetryPersistenceEnabled();
-
-const router = createTelemetryRouter({
-  usePersistence,
-  defaultTenantId: 'DEV',
-  fallbackStore: getTelemetryFallbackStore(),
-  getPersistentStore: () => ensureTelemetryPersistentStore(),
-});
-
-export default router;
+export default telemetryRouter;
